backend/requirements.txt
fastapi
uvicorn
python-multipart
qdrant-client
sentence-transformers
openai
python-dotenv
requests
sqlalchemy
psycopg2-binary
PyJWT
redis
google-auth
resend
slack_sdk
